requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.15.1
python-multipart==0.0.20
